requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
catboost>=1.2.0
xgboost>=2.0.0
optuna>=3.3.0

# Deep Learning (optional)
# torch>=2.0.0
# tensorflow>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Time Series & Financial
statsmodels>=0.14.0
arch>=8.0.0

# Experiment Tracking
# mlflow>=2.7.0
# wandb>=0.15.0

# Utilities
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Kaggle
kaggle>=1.5.16

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
